refactor(engine): Use ListModel generic model directly
